+++ .vscode/settings.json
@@ -1,3 +1,7 @@
 {
-    "editor.formatOnSave": true
+    "editor.formatOnSave": true,
+    "cSpell.words": [
+        "logoicon",
+        "signup"
+    ]
 }
