+++ src/assets/icons/index.ts
@@ -1,0 +1,2 @@
+export { ReactComponent as IcCool } from "./coolicon.svg";
+export { ReactComponent as IcLogo } from "./logoicon.svg";
